requirements.txt
alembic==1.7.5
anyio==3.4.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
appnope==0.1.3
argcomplete==1.12.3
asttokens==2.0.5
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==36.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
Flask==1.1.4
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
ipykernel==6.12.1
ipython==8.2.0
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==7.2.2
jupyter-core==4.9.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.21.4
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.3.4
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prison==0.2.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.1
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-calendarific==1.0.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
rfc3986==1.5.0
rich==10.16.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.1
traitlets==5.1.1
unicodecsv==0.14.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.7.0
